requirements.txt
streamlit==1.23.1
plotly==5.15.0
pandas==1.4.3
numpy==1.23.1
folium==0.14.0
matplotlib==3.5.3
matplolib-inlibe==0.1.6
haversine==2.7.0
streamlit-folium==0.12.0
Pillow==9.4.0
st-pages==0.4.1
plotly-express==0.4.1